Organize linalg rewrite tests according to library
